+++ GeosCore/photolysis_mod.F90
@@ -109,10 +109,10 @@ CONTAINS
     !=======================================================================
 
     ! Initialize
-    RC = GC_SUCCESS
-    notDryRun = ( .not. Input_Opt%DryRun )
-    ErrMsg = ''
-    ThisLoc = ' -> at Init_Photolysis  (in module GeosCore/photolysis_mod.F90)'
+    RC      = GC_SUCCESS
+    ErrMsg  = ''
+    ThisLoc = &
+      ' -> at Init_Photolysis  (in module GeosCore/photolysis_mod.F90)'
 
     ! Set pointers
     GC_Photo_ID => State_Chm%Phot%GC_Photo_ID
@@ -138,10 +138,10 @@ CONTAINS
        ENDIF
     ENDIF
 
-    !--------------------------------------------------------------------
+    !------------------------------------------------------------------------
     ! Read in AOD data even if photolysis disabled
     ! (or just print file name if in dry-run mode)
-    !--------------------------------------------------------------------
+    !------------------------------------------------------------------------
     CALL RD_AOD( Input_Opt, State_Chm, RC )
     IF ( RC /= GC_SUCCESS ) THEN
        ErrMsg = 'Error encountered in FAST-JX routine "RD_AOD"!'
@@ -149,21 +149,25 @@ CONTAINS
        RETURN
     ENDIF
 
+    !------------------------------------------------------------------------
+    ! Exit without doing any computations if we are doing a dry-run
+    !------------------------------------------------------------------------
+    IF ( Input_Opt%DryRun ) RETURN
+
+    !------------------------------------------------------------------------
     ! Compute the required wavelengths in the LUT to calculate requested AOD
-    IF ( notDryRun ) THEN
-       IF (Input_Opt%amIRoot) WRITE(6,*) 'Wavelength optics read successfully'
-       CALL CALC_AOD( Input_Opt, State_Chm, RC )
-    ENDIF
+    !------------------------------------------------------------------------
+    IF (Input_Opt%amIRoot) WRITE(6,*) 'Wavelength optics read successfully'
+    CALL CALC_AOD( Input_Opt, State_Chm, RC )
 
-    !--------------------------------------------------------------------
-    ! Exit if photolysis disabled (zero J-values) or if dry-run
-    !--------------------------------------------------------------------
+    !------------------------------------------------------------------------
+    ! Exit if photolysis disabled (zero J-values)
+    !------------------------------------------------------------------------
     IF ( .NOT. Input_Opt%Do_Photolysis ) RETURN
-    IF ( Input_Opt%DryRun              ) RETURN
 
-    !--------------------------------------------------------------------
+    !------------------------------------------------------------------------
     ! Set up MIEDX array to interpret between GC and FJX aerosol indexing
-    !--------------------------------------------------------------------
+    !------------------------------------------------------------------------
     CALL SET_AER( Input_Opt, State_Chm, RC )
 
     !========================================================================
@@ -392,7 +396,7 @@ CONTAINS
     ENDIF
 
     ! Skip further processing if we are in dry-run mode
-    IF ( notDryRun ) THEN
+    IF (  .not. Input_Opt%DryRun ) THEN
 
        ! Define species IDs
        id_NIT  = IND_('NIT')